refactor: introduce common storage container network node
