+++ refinedstorage2-network/src/test/java/com/refinedmods/refinedstorage2/api/network/impl/node/storage/StorageNetworkNodeProviderImpl.java
@@ -1,12 +1,13 @@
 package com.refinedmods.refinedstorage2.api.network.impl.node.storage;
 
+import com.refinedmods.refinedstorage2.api.network.impl.node.AbstractStorageContainerNetworkNode;
 import com.refinedmods.refinedstorage2.api.storage.Storage;
 
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 
-class StorageNetworkNodeProviderImpl implements StorageNetworkNode.Provider {
+class StorageNetworkNodeProviderImpl implements AbstractStorageContainerNetworkNode.Provider {
     private final Map<Integer, Storage> storages = new HashMap<>();
 
     @Override
